fix(theme): give checkbox/radio icons a fixed size instead of min-content
Bruttal theme's checkbox and radio svg icons were sized with
`width/height: min` (min-content). WebKit/Safari resolves min-content
on a replaced element with no intrinsic size (an svg with only a
viewBox) very differently from Firefox/Chromium, falling back to the
default replaced-element box and rendering the icon at a huge size.
Firefox/Chrome infer a small size from the surrounding layout, which
is why the bug only shows up in Safari.

Also splits BruttalTheme.ts (colors/styles/badges/buttons/forms/text)
into sibling files, mirroring the existing BruttalFonts.ts/
BruttalIcons.ts pattern, since the file already exceeded the repo's
max-lines lint rule before this change and any edit to it tripped the
pre-commit hook. No values changed by the split (locked in by the
existing Bruttal theme snapshot test).
