+++ packages/theme/src/themes/Bruttal/BruttalBadges.ts
@@ -1,0 +1,24 @@
+export const BruttalBadges = {
+  positive: {
+    color: 'action.text.accent.default',
+    bg: 'action.background.accent.active',
+  },
+  negative: {
+    color: 'action.text.negative.default',
+    bg: 'action.background.negative.default',
+  },
+  neutral: {
+    color: 'action.text.primary.default',
+    bg: 'action.background.secondary.default',
+    border: 'sm',
+    borderColor: 'action.border.muted.default',
+  },
+  informative: {
+    color: 'action.text.caution.default',
+    bg: 'action.background.caution.default',
+  },
+  muted: {
+    color: 'action.text.muted.default',
+    bg: 'action.background.muted.default',
+  },
+};
